+++ api/graphql/resolvers/category.resolvers.js
@@ -1,9 +1,10 @@
+const boom = require('@hapi/boom');
 const CategoryService = require('../../services/category.service');
 const service = new CategoryService();
 
 /**
  * Finds all categories in the array of objects
- * @returns {Array} Array with all categories
+ * @returns {array} Array with all categories
  */
 const categories = () => {
 	return service.find();
@@ -12,7 +13,7 @@ const categories = () => {
 /**
  * Finds the category with the provided id
  * @param {id} id - id of the category
- * @returns {Object} Object with the category
+ * @returns {object} Object with the category
  */
 const category = (_, { id }) => {
 	return service.findOne(id);
@@ -20,37 +21,44 @@ const category = (_, { id }) => {
 
 /**
  * Creates a category with the provided data
- * @param {*} data - data of the category
- * @returns {Object} Object with the category created
+ * @param {object} data - data of the category
+ * @returns {object} Object with the category created
  */
-const createCategory = (_, { data }) => {
+const createCategory = async (_, { data }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	return service.create(data);
 };
 
 /**
- * Updates the category with the provided slug
- * @param {slug} slug - slug of the category
- * @param {data} changes - data of the category
- * @returns {Object} Object with the category updated
+ * Updates the category with the provided id
+ * @param {object} params - id and data of the category
+ * @param {string} params.id - id of the category
+ * @param {object} params.data - data of the category
+ * @returns {object} Object with the category updated
  */
-const updateCategory = (_, { slug, data }) => {
+const updateCategory = async (_, { slug, data }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	return service.update(slug, data);
 };
 
 /**
  * Deletes the category with the provided slug
- * @param {slug} slug - slug of the category
- * @returns {Object} Object with the category deleted
+ * @param {string} slug - slug of the category
+ * @returns {object} Object with the category deleted
  */
-const deleteCategory = async (_, { slug }) => {
+const deleteCategory = async (_, { slug }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	await service.delete(slug);
 	return slug;
 };
 
-module.exports = {
-	categories,
-	category,
-	createCategory,
-	updateCategory,
-	deleteCategory,
-};
+module.exports = { categories, category, createCategory, updateCategory, deleteCategory };
