+++ api/graphql/resolvers/payment-method.resolvers.js
@@ -1,3 +1,4 @@
+const boom = require('@hapi/boom');
 const PaymentMethodService = require('../../services/payment-method.service');
 const service = new PaymentMethodService();
 
@@ -12,37 +13,54 @@ const paymentMethods = () => {
 /**
  * Finds the payment method with the provided id
  * @param {id} id - id of the payment method
- * @returns {Object} Object with the payment method
+ * @returns {object} Object with the payment method
  */
-const paymentMethod = (_, { id }) => {
+const paymentMethod = async (_, { id }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	return service.findOne(id);
 };
 
 /**
  * Creates a payment method with the provided data
- * @param {*} data - data of the payment method
- * @returns {Object} Object with the payment method created
+ * @param {object} data - data of the payment method
+ * @returns {object} Object with the payment method created
  */
-const createPaymentMethod = (_, { data }) => {
+const createPaymentMethod = async (_, { data }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	return service.create(data);
 };
 
 /**
- * Updates the payment method with the provided id
- * @param {id} id - id of the payment method
- * @param {data} changes - data of the payment method
- * @returns {Object} Object with the payment method updated
+ * Updates the address with the provided id
+ * @param {object} params - id and data of the address
+ * @param {string} params.id - id of the address
+ * @param {object} params.data - data of the address
+ * @returns {object} Object with the address updated
  */
-const updatePaymentMethod = (_, { id, data }) => {
+const updatePaymentMethod = async (_, { id, data }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	return service.update(id, data);
 };
 
 /**
  * Deletes the payment method with the provided id
- * @param {id} id - id of the payment method
- * @returns {Object} Object with the payment method deleted
+ * @param {string} id - id of the payment method
+ * @returns {object} Object with the payment method deleted
  */
-const deletePaymentMethod = async (_, { id }) => {
+const deletePaymentMethod = async (_, { id }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	await service.delete(id);
 	return id;
 };
