+++ api/graphql/resolvers/user.resolvers.js
@@ -1,9 +1,10 @@
+const boom = require('@hapi/boom');
 const UserService = require('../../services/user.service');
 const service = new UserService();
 
 /**
  * Finds all users in the array of objects
- * @returns {Array} Array with all users
+ * @returns {array} Array with all users
  */
 const users = () => {
 	return service.find();
@@ -11,8 +12,8 @@ const users = () => {
 
 /**
  * Finds the user with the provided username
- * @param {username} username - username of the user
- * @returns {Object} Object with the user
+ * @param {string} username - username of the user
+ * @returns {object} Object with the user
  */
 const user = (_, { username }) => {
 	return service.findByUsername(username);
@@ -20,8 +21,8 @@ const user = (_, { username }) => {
 
 /**
  * Creates a user with the provided data
- * @param {*} data - data of the user
- * @returns {Object} Object with the user created
+ * @param {object} data - data of the user
+ * @returns {object} Object with the user created
  */
 const createUser = (_, { data }) => {
 	return service.create(data);
@@ -29,20 +30,29 @@ const createUser = (_, { data }) => {
 
 /**
  * Updates the user with the provided username
- * @param {username} username - username of the user
- * @param {data} changes - data of the user
- * @returns {Object} Object with the user updated
+ * @param {object} params - username and data of the user
+ * @param {string} params.username - username of the user
+ * @param {object} params.data - data of the user
+ * @returns {object} Object with the user updated
  */
-const updateUser = (_, { username, data }) => {
+const updateUser = async (_, { username, data }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	return service.update(username, data);
 };
 
 /**
  * Deletes the user with the provided username
- * @param {username} username - username of the user
- * @returns {Object} Object with the user deleted
+ * @param {string} params - username of the user
+ * @returns {object} Object with the user deleted
  */
-const deleteUser = async (_, { username }) => {
+const deleteUser = async (_, { username }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+	if (!user) {
+		throw boom.unauthorized('No tienes permiso para realizar esta acción');
+	}
 	await service.delete(username);
 	return username;
 };
